feat: build backend for task creation form
modify Project.js to add new methods (addTaskItem,
removeTaskItem, getTaskItemById) and add a new
array for taskItems assigned directly to the
project without a taskList.

create new functions in projectController.js that
utilize the new methods in Project.js
(addTaskItemToProject and
removeTaskItemFromProject), and modify
viewAllProjects() to now show unassigned tasks
that belong to the project.

modify createTaskItem() in taskItemController.js
to optionally take projectId and taskListIds and
add them to either a taskList or Project depending
on which values were received.

create a new test in index.js to test backend
functionality and log the results in the console.