feat: build frontend for unassigned task items
fix projectFormSubmission() to correctly collect
the id of the currently expanded project.

modify taskFormSubmission() to grab the project id
that a task will be assigned to if not assigned to
a taskList.

create a new file taskGenerator.js with new
functions taskItemElement() and addTasktoUI() to
create a visual representation of the task and
assign it to either a taskList or the project
itself.

add a new hidden styling to style.css to only
display the unassigned task section when something
is added.