feat: send new task list submission to drop down
create new function handleTaskListSubmission()
that captures the new task list title on submit,
validates the input (non-empty), creates and store
a new TaskList object, adds it to the correct
project, adds it as an <option> in the task list
dropdown, and sets it as the selected option.

call the function in appUI.js.