fix: make list dropdown options unique to project
add a new function populateTaskListDropdown() in
taskGenerator.js that builds the dropdown options
each time a dialog is opened, per project, to
avoid having the same options across all projects.

modify the new task button event listener to call
the new function on click in additional to opening
the modal.