feat: make task li elements expand/collapse
modify createTaskItemElement() in taskGenerator.js
to create a color-coded header (based on its
priority level) that displays just its title for
the li task element, and also create a hidden div
element that will drop down on click and contains
all of the task details.

add a slight modification to addTaskToUI() in
taskGenerator.js to allow both the if and else
statements access to the taskElement created when
calling createTaskItemElement().

update css to stylize the li color, header, and
dropdown settings.