feat: add arrow icon, auto collapse on card shrink
add arrow icon to the task item header that
rotates depending on if the dropdown is
open/closed in createTaskItemElement().

modify shrinkCard() in projectCardHandler.js to
automatically collapse all task dropdowns when
closing a project card (this change also
eliminated a bug that caused the modal overlay
effect to disappear when clicking on other cards
if a previous card left a task dropdown open).