feat: add task completion checkbox
modify createTaskItemElement() in taskGenerator.js
to create a new checkbox that toggles the TaskItem
checkbox constructor property, also add an
encompassing div to allow the checkbox to exist
alongside each task element.

add and adjust css to allow the checkbox to appear
next to tasks and give a grey-out effect on
tasks marked "completed".