fix: alter drawer and modal select bar dimensions
adjust the css for the task item details drawer to
prevent long descriptions/notes text from pushing
the header's task arrow off screen when expanded.

remove the height limit on the task drawer to
expand as deep as content goes.

flip the grid-template-columns order on task list
headers to make the arrow stay fixed and allow the
title to fill the rest of the space -- this was
needed due to the above changes altering the
original spacing.

prevent long titles in the new task modal's task
list dropdown box from expanding the modal's
window when too long.