feat: add delete button for tasks and lists
add code to createTaskItemElement() and
createTaskListElement() in taskGenerator.js that
create delete buttons and run the removal
functions found in projectController.js to delete
the items from the DOM (note removing the list
also removes all tasks assigned to that list
without needing extra code for that desired
effect).

add css to style the buttons and make
modifications to existing css to make everything
align properly.