feat: add ability to edit task text
add new dialog modal in template.html for editing
existing task items input fields.

add css styling for the cog button's positioning
inside of the taskDetails dropdown and for the
modal itself.

create new listeners inside appUI.js that listen
for close or submit button clicks, and update task
accordingly if submitted.

add cog icon and listener inside of
createTaskItemElement() in taskGenerator.js and
add a new function that takes the collected data
from the appUI.js listener and updates the
taskItem object in the data storage.

remove toggleCheckbox() method from TaskItem.js as
it was unnecessary and was replaced with a direct
call inside of taskGenerator.js.